requirements.txt
django
djangorestframework
mysqlclient
jsonfield
django-cors-headers
